ci: build bpftool from source on ubuntu
bpftool needs a particular kernel version in order to run.
bpftool is needed to generate vmlinux on the build machine.
This commit ensures that version of bpftool in the container
works in the containerized environment.

Previously had issues because ubuntu-jammy didn't have the
newer linux-tools packages and so any version of bpftool in a
ubuntu jammy container was incompatabile with the host
ubuntu-noble runner.

Signed-off-by: Alan Wandke <adwandk@uwe.nsa.gov>